factor out the landing page/landing component into separate elements
