+++ js/pages/landing.jsx
@@ -1,9 +1,8 @@
 import React, { Component } from 'react';
 import { Container } from 'flux/utils';
 
+import LandingComponent from '../components/landing';
 import { requestActions } from 'actions';
-import AgencyComponentSelector from 'components/agency_component_selector';
-import AgencyComponentPreview from 'components/agency_component_preview';
 import agencyComponentStore from '../stores/agency_component';
 
 
@@ -13,10 +12,9 @@ class LandingPage extends Component {
   }
 
   static calculateState() {
-    const { agencyComponent, agencyComponents, agencies } = agencyComponentStore.getState();
+    const { agencyComponents, agencies } = agencyComponentStore.getState();
     return {
       agencies,
-      agencyComponent,
       agencyComponents,
     };
   }
@@ -33,30 +31,9 @@ class LandingPage extends Component {
 
 
   render() {
-    const agencyChange = (agencyComponent) => {
-      if (agencyComponent.type === 'agency') {
-        // TODO we probably want to make top-level agencies un-selectable?
-        return;
-      }
-
-      requestActions.fetchAgencyComponent(agencyComponent.id)
-        .then(requestActions.receiveAgencyComponent);
-    };
-
     const { agencies, agencyComponents } = this.state;
     return (
-      <div className="usa-grid">
-        <h2>Select the agency that you’d like to send a FOIA request to</h2>
-        <AgencyComponentSelector
-          agencies={agencies}
-          agencyComponents={agencyComponents}
-          onAgencyChange={agencyChange}
-        />
-        {
-          this.state.agencyComponent.id &&
-          <AgencyComponentPreview agencyComponent={this.state.agencyComponent.toJS()} />
-        }
-      </div>
+      <LandingComponent agencies={agencies} agencyComponents={agencyComponents} />
     );
   }
 }
